REFACTOR: Moved skill related functions into skill directory
